refactoring: moved useful functions to dif library
