+++ src/c/utils.c
@@ -1,0 +1,15 @@
+#include "utils.h"
+#include "helper.h"
+
+int randomIntRange(int min, int max)
+{
+    int value = randomInt();
+    return (value % (max - min + 1)) + min;
+}
+
+
+int map(int x, int in_min, int in_max, int out_min, int out_max)
+{
+    //vgl Arduino map() https://www.arduino.cc/reference/en/language/functions/math/map/
+    return (x - in_min) * (out_max - out_min) / (in_max - in_min) + out_min;
+}
