+++ src/c/utils.h
@@ -1,0 +1,7 @@
+#ifndef UTILS_H
+#define UTILS_H
+
+int randomIntRange(int min, int max);
+int map(int x, int in_min, int in_max, int out_min, int out_max);
+
+#endif
